+++ src/powerplatform/mockConnector.ts
@@ -6,6 +6,8 @@ import { Logger } from "../utils/Logger";
  *
  * @param page - The Playwright Page object.
  * @param connectorId - The ID of the connector to mock.
+ * @param method - The HTTP method to mock. By default, it will mock a GET request.
+ * @param status - The status code to return in the mocked response. By default, it will return 200.
  * @param data - The data to return in the mocked response.
  * @param options - Optional settings for the route.
  * @param options.times - How often the route should be used. By default, it will be used every time.
@@ -16,6 +18,8 @@ export const mockConnector = async (
   page: Page,
   connectorId: string,
   data: any,
+  method: string = "GET",
+  status: number = 200,
   options?: {
     logging?: boolean;
     /**
@@ -30,18 +34,20 @@ export const mockConnector = async (
   return await page.route(
     "**/invoke",
     async (route) => {
-      const method = route.request().method();
-      if (method !== "POST") {
+      const apiMethod = route.request().method();
+      if (apiMethod !== "POST") {
         return route.continue();
       }
 
       const headers = route.request().headers();
       const requestUrl = headers["x-ms-request-url"];
+      const requestMethod = headers["x-ms-request-method"];
 
       if (logging) {
         Logger(
           `mockConnector: Checking connector ID in request URL...`,
-          requestUrl
+          requestUrl,
+          requestMethod
         );
       }
 
@@ -49,12 +55,18 @@ export const mockConnector = async (
         return route.continue();
       }
 
+      if (!requestMethod || requestMethod !== method) {
+        return route.continue();
+      }
+
       if (logging) {
-        Logger(`mockConnector: Mocking response for ${connectorId}...`);
+        Logger(
+          `mockConnector: Mocking response for (${requestMethod}) ${connectorId}...`
+        );
       }
 
       route.fulfill({
-        status: 200,
+        status: status,
         contentType: "application/json",
         body: JSON.stringify(data),
       });