+++ src/powerplatform/waitForConnectorResponse.ts
@@ -6,6 +6,7 @@ import { Logger } from "../utils/Logger";
  *
  * @param page - The Playwright Page object to interact with.
  * @param connectorId - The ID of the connector to wait for a response from.
+ * @param method - The HTTP method to wait for. By default, it will wait for a GET request.
  * @param options - Optional settings for waiting for the response.
  * @param options.logging - Whether to log the waiting process.
  * @param options.timeout - The maximum time to wait for the response, in milliseconds.
@@ -14,6 +15,7 @@ import { Logger } from "../utils/Logger";
 export const waitForConnectorResponse = async (
   page: Page,
   connectorId: string,
+  method: string = "GET",
   options?: {
     logging?: boolean;
     timeout?: number;
@@ -31,11 +33,13 @@ export const waitForConnectorResponse = async (
 
     const headers = response.request().headers();
     const requestUrl = headers["x-ms-request-url"];
+    const requestMethod = headers["x-ms-request-method"];
 
     if (logging) {
       Logger(
         `waitForConnectorResponse: Checking connector ID in request URL...`,
-        requestUrl
+        requestUrl,
+        requestMethod
       );
     }
 
@@ -43,8 +47,14 @@ export const waitForConnectorResponse = async (
       return false;
     }
 
+    if (!requestMethod || requestMethod !== method) {
+      return false;
+    }
+
     if (logging) {
-      Logger(`waitForConnectorResponse: Response from ${connectorId} found!`);
+      Logger(
+        `waitForConnectorResponse: Response from  (${requestMethod}) ${connectorId} found!`
+      );
     }
 
     return true;